requirements.txt
Django<=1.8
django-debug-toolbar==1.3.0
django-filter==0.9.2
django-filters==0.1.0
django-redis==3.8.3
djangorestframework==3.1.1
drf-extensions==0.2.7
ipython==3.1.0
mysqlclient==1.3.6
redis==2.10.3
sqlparse==0.1.15